0.13.2.1

+++ ingestion/src/metadata/ingestion/bulksink/metadata_usage.py
@@ -8,11 +8,21 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+"""
+BulkSink class used for Usage workflows.
+
+It sends Table queries and usage counts to Entities,
+as well as populating JOIN information.
+
+It picks up the information from reading the files
+produced by the stage. At the end, the path is removed.
+"""
 import json
 import os
 import shutil
 import traceback
 from datetime import datetime
+from pathlib import Path
 from typing import List, Optional
 
 from pydantic import ValidationError
@@ -22,7 +32,6 @@ from metadata.generated.schema.entity.data.database import Database
 from metadata.generated.schema.entity.data.table import (
     ColumnJoins,
     JoinedWith,
-    SqlQuery,
     Table,
     TableJoins,
 )
@@ -31,17 +40,16 @@ from metadata.generated.schema.entity.services.connections.metadata.openMetadata
 )
 from metadata.generated.schema.type.tableUsageCount import TableColumn, TableUsageCount
 from metadata.generated.schema.type.usageRequest import UsageRequest
-from metadata.ingestion.api.bulk_sink import BulkSink, BulkSinkStatus
+from metadata.ingestion.api.bulk_sink import BulkSink
+from metadata.ingestion.lineage.sql_lineage import (
+    get_column_fqn,
+    get_table_entities_from_query,
+)
 from metadata.ingestion.ometa.client import APIError
 from metadata.ingestion.ometa.ometa_api import OpenMetadata
 from metadata.utils import fqn
+from metadata.utils.constants import UTF_8
 from metadata.utils.logger import ingestion_logger
-from metadata.utils.lru_cache import LRUCache
-from metadata.utils.sql_lineage import (
-    get_column_fqn,
-    get_lineage_by_query,
-    get_table_entities_from_query,
-)
 
 logger = ingestion_logger()
 
@@ -53,6 +61,13 @@ class MetadataUsageSinkConfig(ConfigModel):
 
 
 class MetadataUsageBulkSink(BulkSink):
+    """
+    BulkSink implementation to send:
+    - table usage
+    - table queries
+    - frequent joins
+    """
+
     config: MetadataUsageSinkConfig
 
     def __init__(
@@ -60,13 +75,12 @@ class MetadataUsageBulkSink(BulkSink):
         config: MetadataUsageSinkConfig,
         metadata_config: OpenMetadataConnection,
     ):
-
+        super().__init__()
         self.config = config
         self.metadata_config = metadata_config
         self.service_name = None
         self.wrote_something = False
         self.metadata = OpenMetadata(self.metadata_config)
-        self.status = BulkSinkStatus()
         self.table_join_dict = {}
         self.table_usage_map = {}
         self.today = datetime.today().strftime("%Y-%m-%d")
@@ -75,35 +89,6 @@ class MetadataUsageBulkSink(BulkSink):
     def create(cls, config_dict: dict, metadata_config: OpenMetadataConnection):
         config = MetadataUsageSinkConfig.parse_obj(config_dict)
         return cls(config, metadata_config)
-
-    def ingest_sql_queries_lineage(
-        self, queries: List[SqlQuery], database_name: str, schema_name: str
-    ) -> None:
-        """
-        Method to ingest lineage by sql queries
-        """
-
-        create_or_insert_queries = [
-            query.query
-            for query in queries
-            if "create" in query.query.lower() or "insert" in query.query.lower()
-        ]
-        seen_queries = LRUCache(LRU_CACHE_SIZE)
-
-        for query in create_or_insert_queries:
-            if query in seen_queries:
-                continue
-            lineages = get_lineage_by_query(
-                self.metadata,
-                query=query,
-                service_name=self.service_name,
-                database_name=database_name,
-                schema_name=schema_name,
-            )
-            for lineage in lineages or []:
-                created_lineage = self.metadata.add_lineage(lineage)
-                logger.info(f"Successfully added Lineage {created_lineage}")
-            seen_queries.put(query, None)  # None because it really doesn't matter.
 
     def __populate_table_usage_map(
         self, table_entity: Table, table_usage: TableUsageCount
@@ -116,7 +101,6 @@ class MetadataUsageBulkSink(BulkSink):
             self.table_usage_map[table_entity.id.__root__] = {
                 "table_entity": table_entity,
                 "usage_count": table_usage.count,
-                "sql_queries": table_usage.sqlQueries,
                 "usage_date": table_usage.date,
                 "database": table_usage.databaseName,
                 "database_schema": table_usage.databaseSchema,
@@ -125,53 +109,42 @@ class MetadataUsageBulkSink(BulkSink):
             self.table_usage_map[table_entity.id.__root__][
                 "usage_count"
             ] += table_usage.count
-            self.table_usage_map[table_entity.id.__root__]["sql_queries"].extend(
-                table_usage.sqlQueries
-            )
 
     def __publish_usage_records(self) -> None:
         """
-        Method to publish SQL Queries, Table Usage & Lineage
+        Method to publish SQL Queries, Table Usage
         """
         for _, value_dict in self.table_usage_map.items():
             table_usage_request = None
             try:
                 table_usage_request = UsageRequest(
-                    date=value_dict["usage_date"], count=value_dict["usage_count"]
-                )
-                self.metadata.ingest_table_queries_data(
-                    table=value_dict["table_entity"],
-                    table_queries=value_dict["sql_queries"],
-                )
-                self.ingest_sql_queries_lineage(
-                    queries=value_dict["sql_queries"],
-                    database_name=value_dict["database"],
-                    schema_name=value_dict["database_schema"],
+                    date=datetime.fromtimestamp(int(value_dict["usage_date"])).strftime(
+                        "%Y-%m-%d"
+                    ),
+                    count=value_dict["usage_count"],
                 )
                 self.metadata.publish_table_usage(
                     value_dict["table_entity"], table_usage_request
                 )
                 logger.info(
-                    "Successfully table usage published for {}".format(
-                        value_dict["table_entity"].name.__root__
-                    )
+                    f"Successfully table usage published for {value_dict['table_entity'].fullyQualifiedName.__root__}"
                 )
                 self.status.records_written(
-                    "Table: {}".format(value_dict["table_entity"].name.__root__)
+                    f"Table: {value_dict['table_entity'].fullyQualifiedName.__root__}"
                 )
             except ValidationError as err:
-                logger.error(
-                    f"Cannot construct UsageRequest from {value_dict['table_entity']} - {err}"
+                logger.debug(traceback.format_exc())
+                logger.warning(
+                    f"Cannot construct UsageRequest from {value_dict['table_entity']}: {err}"
                 )
-            except Exception as err:
+            except Exception as exc:
+                logger.debug(traceback.format_exc())
+                logger.warning(
+                    f"Failed to update usage for {value_dict['table_entity'].fullyQualifiedName.__root__} :{exc}"
+                )
                 self.status.failures.append(table_usage_request)
-                logger.error(
-                    "Failed to update usage for {} {}".format(
-                        value_dict["table_entity"].name.__root__, err
-                    )
-                )
                 self.status.failures.append(
-                    "Table: {}".format(value_dict["table_entity"].name.__root__)
+                    f"Table: {value_dict['table_entity'].fullyQualifiedName.__root__}"
                 )
 
     def iterate_files(self):
@@ -184,7 +157,7 @@ class MetadataUsageBulkSink(BulkSink):
                 full_file_name = os.path.join(self.config.filename, filename)
                 if not os.path.isfile(full_file_name):
                     continue
-                with open(full_file_name) as file:
+                with open(full_file_name, encoding=UTF_8) as file:
                     yield file
 
     # Check here how to properly pick up ES and/or table query data
@@ -205,11 +178,11 @@ class MetadataUsageBulkSink(BulkSink):
                         database_schema=table_usage.databaseSchema,
                         table_name=table_usage.table,
                     )
-                except Exception as err:
-                    logger.error(
-                        f"Cannot get table entities from query table {table_usage.table} - {err}"
-                    )
+                except Exception as exc:
                     logger.debug(traceback.format_exc())
+                    logger.warning(
+                        f"Cannot get table entities from query table {table_usage.table}: {exc}"
+                    )
 
                 if not table_entities:
                     logger.warning(
@@ -217,51 +190,58 @@ class MetadataUsageBulkSink(BulkSink):
                     )
                     continue
 
-                for table_entity in table_entities:
-                    if table_entity is not None:
-                        table_join_request = None
-                        try:
-                            self.__populate_table_usage_map(
-                                table_usage=table_usage, table_entity=table_entity
-                            )
-                            table_join_request = self.__get_table_joins(
-                                table_entity=table_entity, table_usage=table_usage
-                            )
-                            logger.debug(
-                                "table join request {}".format(table_join_request)
-                            )
-
-                            if (
-                                table_join_request is not None
-                                and len(table_join_request.columnJoins) > 0
-                            ):
-                                self.metadata.publish_frequently_joined_with(
-                                    table_entity, table_join_request
-                                )
-                        except APIError as err:
-                            self.status.failures.append(table_join_request)
-                            logger.error(
-                                "Failed to update query join for {}, {}".format(
-                                    table_usage.table, err
-                                )
-                            )
-                        except Exception as err:
-                            logger.error(
-                                f"Error getting usage and join information for {table_entity.name.__root__} - {err}"
-                            )
-                            logger.debug(traceback.format_exc())
-                    else:
-                        logger.warning(
-                            f"Could not fetch table {table_usage.databaseName}.{table_usage.databaseSchema}.{table_usage.table}"
-                        )
-                        self.status.warnings.append(f"Table: {table_usage.table}")
+                self.get_table_usage_and_joins(table_entities, table_usage)
 
             self.__publish_usage_records()
-        try:
-            self.metadata.compute_percentile(Table, self.today)
-            self.metadata.compute_percentile(Database, self.today)
-        except APIError:
-            logger.error("Failed to publish compute.percentile")
+
+    def get_table_usage_and_joins(
+        self, table_entities: List[Table], table_usage: TableUsageCount
+    ):
+        """
+        For the list of tables, compute usage with already existing seen
+        tables and publish the join information.
+        """
+        for table_entity in table_entities:
+            if table_entity is not None:
+                table_join_request = None
+                try:
+                    self.__populate_table_usage_map(
+                        table_usage=table_usage, table_entity=table_entity
+                    )
+                    table_join_request = self.__get_table_joins(
+                        table_entity=table_entity, table_usage=table_usage
+                    )
+                    logger.debug(f"table join request {table_join_request}")
+
+                    if (
+                        table_join_request is not None
+                        and len(table_join_request.columnJoins) > 0
+                    ):
+                        self.metadata.publish_frequently_joined_with(
+                            table_entity, table_join_request
+                        )
+
+                    if table_usage.sqlQueries:
+                        self.metadata.ingest_entity_queries_data(
+                            entity=table_entity, queries=table_usage.sqlQueries
+                        )
+                except APIError as err:
+                    logger.debug(traceback.format_exc())
+                    logger.warning(
+                        f"Failed to update query join for {table_usage.table}: {err}"
+                    )
+                    self.status.failures.append(table_join_request)
+                except Exception as exc:
+                    logger.debug(traceback.format_exc())
+                    logger.warning(
+                        f"Error getting usage and join information for {table_entity.name.__root__}: {exc}"
+                    )
+            else:
+                logger.warning(
+                    "Could not fetch table"
+                    f" {table_usage.databaseName}.{table_usage.databaseSchema}.{table_usage.table}"
+                )
+                self.status.warnings.append(f"Table: {table_usage.table}")
 
     def __get_table_joins(
         self, table_entity: Table, table_usage: TableUsageCount
@@ -278,7 +258,7 @@ class MetadataUsageBulkSink(BulkSink):
             if column_join.tableColumn is None or len(column_join.joinedWith) == 0:
                 continue
 
-            if column_join.tableColumn.column in column_joins_dict.keys():
+            if column_join.tableColumn.column in column_joins_dict:
                 joined_with = column_joins_dict[column_join.tableColumn.column]
             else:
                 column_joins_dict[column_join.tableColumn.column] = {}
@@ -334,9 +314,14 @@ class MetadataUsageBulkSink(BulkSink):
         for table_entity in table_entities:
             return get_column_fqn(table_entity=table_entity, column=table_column.column)
 
-    def get_status(self):
-        return self.status
-
     def close(self):
-        shutil.rmtree(self.config.filename)
+        if Path(self.config.filename).exists():
+            shutil.rmtree(self.config.filename)
+        try:
+            self.metadata.compute_percentile(Table, self.today)
+            self.metadata.compute_percentile(Database, self.today)
+        except APIError as err:
+            logger.debug(traceback.format_exc())
+            logger.warning(f"Failed to publish compute.percentile: {err}")
+
         self.metadata.close()
